fix: Firefox Connection Reset on 413 and Refactor Payload Limits (#125)
* network: implement connection draining to fix Firefox RST on 413

* config: upgrade numeric storage to size_t for large file support

* config: refactor ClientMaxBodySize to support units and robust validation

+++ inc/config/arguments/Integer.hpp
@@ -6,15 +6,15 @@ namespace config {
 
 class Integer : public IArgument {
 public:
-    explicit Integer(int val);
+    explicit Integer(size_t val);
     std::string evaluate(http::Request const &) const;
-    int getIntValue() const;
+    size_t getIntValue() const;
     ArgumentType getType() const { return ARG_INTEGER; }
     std::string getRawValue() const { return sValue_; }
     IArgument *clone() const { return new Integer(*this); }
 
 private:
-    int iValue_;
+    size_t iValue_;
     std::string sValue_;
 };
