+++ src/config/arguments/Integer.cpp
@@ -3,8 +3,8 @@
 
 namespace config {
 
-Integer::Integer(int val) : iValue_(val), sValue_(utils::toString(val)) {}
+Integer::Integer(size_t val) : iValue_(val), sValue_(utils::toString(val)) {}
 std::string Integer::evaluate(http::Request const &) const { return sValue_; }
-int Integer::getIntValue() const { return iValue_; }
+size_t Integer::getIntValue() const { return iValue_; }
 
 } // namespace config
